Rename autossugest-text.tsx to autosuggest-text.tsx to correct the typo
The name was misspelled and inconsistent.
